Completed the header
Wow, web development is harder than i thought.
It's really difficult, in the beginning i thought it was as simple as drag and dropping stuff on a webpage, but like, it's really a whole new theory.
Now i know why it's called software engineering sometimes, because it really has to do with engineering.
Anyways, yeah. After compelting some of the additional reading materials on flexbox, i managed to learn new techniques that i could hone for this project.

Alhamdulilah.