If you look at it at half a window, at 100% scaling. It looks good
But, if you adjust the scaling, zoom in, zoom out, etc... it does not look good.
So yeah, in the future i'll work on that, even in full screen it looks kinda bad, so i've gotta work on that.
Coz on my right side of the screen was the sample, then on the left side it was the actual index.html
So in full screen it looks a little awkward...

come back to this in the future, rn quite sian of html and css alr, wanna learn some new stuff.
Inshallah will come back to this in the future, coz this is also very important, but just wanna take a break.